requirements.txt
arxiv
langchain-community
load_dotenv
wikipedia
faiss-cpu
langchain_ollama
langchain
langchain_groq